test(edit): Shift parse tests to toml_edit
