+++ crates/toml/tests/testsuite/main.rs
@@ -1,15 +1,12 @@
 #![recursion_limit = "256"]
 
-mod datetime;
 mod de_errors;
 mod display;
 mod display_tricky;
 mod enum_external_deserialize;
 mod float;
 mod formatting;
-mod invalid_misc;
 mod macros;
-mod parser;
 mod pretty;
 mod serde;
 mod spanned;
