+++ crates/toml_edit/tests/testsuite/main.rs
@@ -1,6 +1,7 @@
 #![recursion_limit = "256"]
 
 mod convert;
+mod datetime;
 mod edit;
 mod invalid;
 mod parse;
